Subsume browse_everythingable into base_resource_controller
